+++ src/BlazorWebFormsComponents/Button.razor
@@ -1,1 +1,3 @@
-<button @onclick="Click" OnClientClick="@OnClientClick" >@Text</button>
+@inherits ButtonBaseComponent
+
+<button @onclick="Click" OnClientClick="@OnClientClick">@Text</button>
